Moves guts of the code to addon directory. App imports-exports the addons.
