+++ app/components/linkedin-share.js
@@ -1,55 +1,2 @@
-import Ember from 'ember';
-
-export default Ember.Component.extend({
-  socialApiClient: null, // injected
-  tagName: 'div', // set tagName to 'a' in handlebars to use your own css/content
-                  // instead of the standard Linkedin share button UI
-  isCustomLink: Ember.computed.equal('tagName','a'),
-  useLinkedinUi: Ember.computed.not('isCustomLink'),
-
-  count: null, //can be 'top' or 'right'
-  url: null, // Defaults to current url
-  createLinkedinShareButton: Ember.on('didInsertElement', function() {
-    var self = this;
-    this.socialApiClient.load().then(function(IN) {
-      self.IN = IN;
-      self.shareHandlerName = IN.shareHandlerName;
-      if (self._state !== 'inDOM') { return; }
-      if (self.get('useLinkedinUi')) {
-        var attrs = [];
-        var url = self.get('url');
-        var count = self.get('count');
-        if (url) {
-          attrs.push('data-url="' + url + '"');
-        }
-        if(count) {
-          attrs.push('data-counter="' + count +'"');
-        }
-        attrs.push('data-onsuccess="' + self.shareHandlerName + '"');
-        self.$().html('<script type="IN/Share" ' + attrs.join(' ') + '></script>');
-        IN.parse(self.get('element'));
-      } else {
-        self.$().attr('href', '#');
-      }
-    });
-  }),
-
-  showShareDialog: Ember.on('click', function(e){
-    var self = this;
-    this.socialApiClient.clicked(this.get('url') || window.location.href);
-    if (this.get('useLinkedinUi')) { return; }
-    function showDialog(IN) {
-      IN.UI.Share().params({
-        url: self.get('url')
-      }).place().success(window[self.shareHandlerName]);
-    }
-    if (this.IN) {
-      showDialog(this.IN);
-    } else {
-      this.socialApiClient.load().then(function(IN) {
-        showDialog(IN);
-      });
-    }
-    e.preventDefault();
-  })
-});
+import LinkedinShareComponent from 'ember-social/components/linkedin-share';
+export default LinkedinShareComponent;
